refactor: move extract_program_and_entry_points_by_type into serde
