+++ crates/storage/src/contract.rs
@@ -1,8 +1,8 @@
-use crate::state::class_hash::extract_program_and_entry_points_by_type;
 use crate::types::CompressedContract;
 use anyhow::Context;
 use flate2::{write::GzEncoder, Compression};
 use pathfinder_common::{ClassHash, ContractAddress, ContractClass, StarknetBlockHash};
+use pathfinder_serde::extract_program_and_entry_points_by_type;
 use rusqlite::{named_params, Connection, OptionalExtension, Transaction};
 
 /// Stores StarkNet contract information, specifically a contract's
